+++ .changeset/wet-pans-rush.md
@@ -1,5 +1,6 @@
 ---
 '@2digits/prettier-config': minor
+'@2digits/eslint-config': minor
 '@2digits/eslint-plugin': minor
 '@2digits/constants': minor
 ---
