+++ pnpm-workspace.yaml
@@ -93,7 +93,6 @@ catalog:
   turbo: 2.5.4
   typescript: 5.8.3
   typescript-eslint: 8.33.0
-  unbuild: 3.5.0
   vitest: 3.1.4
   yaml-eslint-parser: 1.3.0
 onlyBuiltDependencies:
